i cry
apparently some blocks have blockstate ingame that arent part of the model file (EG powered for doors, or persistent and distance for leaves), so i'll have to give up on this method and check through the json, ignoring values that dont exist in one or the other. i'll probably just make it check one layer and not recursively tho, since for this recursive isn't necesary.
also added a lot of stuff idk if that was in the last update